Change return type of calc_len() family back to int from unsigned long.
svn path=/trunk/yasm/; revision=448

+++ libyasm/bytecode.h
@@ -63,8 +63,8 @@ void bc_print(FILE *f, const bytecode *bc);
  * resolve_label is the function used to determine the value (offset) of a
  *  in-file label (eg, not an EXTERN variable, which is indeterminate).
  */
-unsigned long bc_calc_len(bytecode *bc,
-			  intnum *(*resolve_label) (symrec *sym));
+int bc_calc_len(bytecode *bc,
+		intnum *(*resolve_label) (symrec *sym));
 
 /* void bcs_initialize(bytecodehead *headp); */
 #define	bcs_initialize(headp)	STAILQ_INIT(headp)
